+++ src/blender_mcp/materials/__init__.py
@@ -1,0 +1,8 @@
+from .spec import build_material_spec, _build_spec_from_keys
+from .blender_create import create_material_in_blender
+
+__all__ = [
+    "build_material_spec",
+    "_build_spec_from_keys",
+    "create_material_in_blender",
+]